Map Grid Lines are working quite well.  Fixed a few bugs which showed up as missing lines in low res mode.  Set the geoBoundary steps = 1 (so that it doesn't skip any boundary pixels).  This works fine to have steps in the interpolation sense, but it causes corners to be missed and hence lines near the border to be missed.  Still need the preferences GUI updated.  There is a half-pixel offset in some of the line endings.  It's harmless but unless you really zoom in or you are clearly seeing the pixels you'll never know.  Right now the grid lines start at the center of the border pixels.

+++ beam-core/src/main/java/org/esa/beam/glayer/GraticuleLayer.java
@@ -86,8 +86,8 @@ public class GraticuleLayer extends Layer {
         if (graticule == null) {
             graticule = Graticule.create(raster,
                     getNumGridLines(),
-                    (float) getResLat(),
-                    (float) getResLon());
+                    getResLat(),
+                    getResLon());
         }
         if (graticule != null) {
